Refactoring pcep.go
